+++ tests/test_post.py
@@ -1,11 +1,15 @@
 import testlib
-from testlib import py3k
+from testlib import py3k, md5sum
 import urlfetch
 
 import unittest
 import json
 import random
 
+
+import os
+here = os.path.dirname(os.path.abspath(__file__))
+path = lambda *p: os.path.join(here, *p)
 
 class PostTest(unittest.TestCase):
 
@@ -48,7 +52,7 @@ class PostTest(unittest.TestCase):
         self.assertEqual(o['query_string'], query_string)
 
     def test_file_upload(self):
-        content = open('test.file', 'rb').read()
+        content = open(path('test.file'), 'rb').read()
         files = {}
         files['test.file1'] = (b'test.file1', b'test.file', content)
         #files[b'test.file2'] = (b'test.file2', b'', content)
@@ -58,9 +62,9 @@ class PostTest(unittest.TestCase):
         r = urlfetch.post(
                 testlib.test_server_host,
                 files = {
-                    'test.file1' : open('test.file'),
+                    'test.file1' : open(path('test.file')),
                     #'test.file2' : content,
-                    'test.file3' : ('wangxiaobo', open('test.file')),
+                    'test.file3' : ('wangxiaobo', open(path('test.file'))),
                     'test.file4' : ('wangtwo', content)
                 },
             )
@@ -71,12 +75,14 @@ class PostTest(unittest.TestCase):
         self.assertEqual(sorted(o['files'].keys()), sorted(files.keys()))
 
         for i in files:
-            for j in range(3):
-                self.assertEqual(o['files'][i][j].encode('utf-8'), files[i][j])
+            self.assertEqual(o['files'][i][0].encode('utf-8'), files[i][0])
+            self.assertEqual(o['files'][i][1].encode('utf-8'), files[i][1])
+            self.assertEqual(o['files'][i][2].encode('utf-8'), md5sum(files[i][2]))
+            
 
 
     def test_one_file_upload(self):
-        content = open('test.file', 'rb').read()
+        content = open(path('test.file'), 'rb').read()
         files = {'test.file': (b'test.file', b'test.file', content)}
 
         r = urlfetch.post(
@@ -91,11 +97,12 @@ class PostTest(unittest.TestCase):
         self.assertEqual(o['method'], 'POST')
         self.assertEqual(sorted(o['files'].keys()), sorted(files.keys()))
         for i in files:
-            for j in range(3):
-                self.assertEqual(o['files'][i][j].encode('utf8'), files[i][j])
+            self.assertEqual(o['files'][i][0].encode('utf8'), files[i][0])
+            self.assertEqual(o['files'][i][1].encode('utf8'), files[i][1])
+            self.assertEqual(o['files'][i][2].encode('utf8'), md5sum(files[i][2]))
 
     def test_one_file_upload_gbk(self):
-        content = open('test.file.gbk', 'rb').read()
+        content = open(path('test.file.gbk'), 'rb').read()
         files = {'test.file': (b'test.file', b'test.file', content)}
 
         r = urlfetch.post(
@@ -110,8 +117,9 @@ class PostTest(unittest.TestCase):
         self.assertEqual(o['method'], 'POST')
         self.assertEqual(sorted(o['files'].keys()), sorted(files.keys()))
         for i in files:
-            for j in range(3):
-                self.assertEqual(o['files'][i][j].encode('gb2312'), files[i][j])
+            self.assertEqual(o['files'][i][0].encode('gbk'), files[i][0])
+            self.assertEqual(o['files'][i][1].encode('gbk'), files[i][1])
+            self.assertEqual(o['files'][i][2].encode('gbk'), md5sum(files[i][2]))
 
 if __name__ == '__main__':
     unittest.main()